+++ test/nn/conv/test_simple_conv.py
@@ -36,15 +36,12 @@ def test_simple_conv(aggr, combine_root):
         assert torch.allclose(conv(x1, adj2.t()), out)
 
     if is_full_test():
-        t = '(Tensor, Tensor, OptTensor, Size) -> Tensor'
-        jit = torch.jit.script(conv.jittable(t))
+        jit = torch.jit.script(conv.jittable())
         assert torch.allclose(jit(x1, edge_index), out)
         assert torch.allclose(jit(x1, edge_index, size=(4, 4)), out)
 
-    if is_full_test() and torch_geometric.typing.WITH_TORCH_SPARSE:
-        t = '(Tensor, SparseTensor, OptTensor, Size) -> Tensor'
-        jit = torch.jit.script(conv.jittable(t))
-        assert torch.allclose(jit(x1, adj2.t()), out)
+        if torch_geometric.typing.WITH_TORCH_SPARSE:
+            assert torch.allclose(jit(x1, adj2.t()), out)
 
     # Test bipartite message passing:
     if combine_root != 'self_loop':