+++ test/nn/conv/test_wl_conv_continuous.py
@@ -21,14 +21,10 @@ def test_wl_conv():
         assert torch.allclose(conv(x, adj.t()), out)
 
     if is_full_test():
-        t = '(Tensor, Tensor, OptTensor, Size) -> Tensor'
-        jit = torch.jit.script(conv.jittable(t))
+        jit = torch.jit.script(conv.jittable())
         assert torch.allclose(jit(x, edge_index), out)
-
-    if is_full_test() and torch_geometric.typing.WITH_TORCH_SPARSE:
-        t = '(Tensor, SparseTensor, OptTensor, Size) -> Tensor'
-        jit = torch.jit.script(conv.jittable(t))
-        assert torch.allclose(jit(x, adj.t()), out, atol=1e-6)
+        if torch_geometric.typing.WITH_TORCH_SPARSE:
+            assert torch.allclose(jit(x, adj.t()), out, atol=1e-6)
 
     # Test bipartite message passing:
     x1 = torch.randn(4, 8)
@@ -48,14 +44,11 @@ def test_wl_conv():
         assert torch.allclose(conv((x1, x2), adj.t()), out2)
 
     if is_full_test():
-        t = '(OptPairTensor, Tensor, OptTensor, Size) -> Tensor'
-        jit = torch.jit.script(conv.jittable(t))
+        jit = torch.jit.script(conv.jittable())
         assert torch.allclose(
             jit((x1, None), edge_index, edge_weight, size=(4, 2)), out1)
         assert torch.allclose(jit((x1, x2), edge_index, edge_weight), out2)
 
-    if is_full_test() and torch_geometric.typing.WITH_TORCH_SPARSE:
-        t = '(OptPairTensor, SparseTensor, OptTensor, Size) -> Tensor'
-        jit = torch.jit.script(conv.jittable(t))
-        assert torch.allclose(jit((x1, None), adj.t()), out1, atol=1e-6)
-        assert torch.allclose(jit((x1, x2), adj.t()), out2, atol=1e-6)
+        if torch_geometric.typing.WITH_TORCH_SPARSE:
+            assert torch.allclose(jit((x1, None), adj.t()), out1, atol=1e-6)
+            assert torch.allclose(jit((x1, x2), adj.t()), out2, atol=1e-6)