+++ torch_geometric/datasets/graph_generator/__init__.py
@@ -2,10 +2,12 @@ from .base import GraphGenerator
 from .ba_graph import BAGraph
 from .er_graph import ERGraph
 from .grid_graph import GridGraph
+from .tree_graph import TreeGraph
 
 __all__ = classes = [
     'GraphGenerator',
     'BAGraph',
     'ERGraph',
     'GridGraph',
+    'TreeGraph',
 ]