+++ torch_geometric/nn/conv/wl_conv_continuous.py
@@ -54,7 +54,7 @@ class WLConvContinuous(MessagePassing):
     ) -> Tensor:
 
         if isinstance(x, Tensor):
-            x: OptPairTensor = (x, x)
+            x = (x, x)
 
         # propagate_type: (x: OptPairTensor, edge_weight: OptTensor)
         out = self.propagate(edge_index, x=x, edge_weight=edge_weight,